+++ tests/testthat/test-backend-postgres.R
@@ -23,7 +23,7 @@ test_that("custom stringr functions translated correctly", {
   trans <- function(x) {
     translate_sql(!!enquo(x), con = simulate_postgres())
   }
-  expect_equal(trans(str_replace(x, y, z)), sql("REGEXP_REPLACE(`x`, `y`, `z`)"))
+  expect_equal(trans(str_replace_all(x, y, z)), sql("REGEXP_REPLACE(`x`, `y`, `z`)"))
 })
 
 test_that("two variable aggregates are translated correctly", {